+++ Material.Blazor/Components/Menu/MBMenu.ts
@@ -4,10 +4,11 @@ export function init(elem, dotNetObject) {
     elem._menu = MDCMenu.attachTo(elem);
 
     return new Promise(() => {
-        elem._menu.foundation.handleItemAction = () => {
-            elem._menu.open = false;
+        const closedCallback = () => {
             dotNetObject.invokeMethodAsync('NotifyClosedAsync');
         };
+
+        elem._menu.listen('MDCMenuSurface:closed', closedCallback);
     });
 }